+++ FXLabel/FXLabel.h
@@ -1,7 +1,7 @@
 //
 //  FXLabel.h
 //
-//  Version 1.1
+//  Version 1.2
 //
 //  Created by Nick Lockwood on 20/08/2011.
 //  Copyright 2011 Charcoal Design. All rights reserved.
@@ -39,5 +39,6 @@
 @property (nonatomic, retain) UIColor *gradientEndColor;
 @property (nonatomic, assign) CGPoint gradientStartPoint;
 @property (nonatomic, assign) CGPoint gradientEndPoint;
+@property (nonatomic, assign) BOOL oversample;
 
 @end